requirements.txt
python-telegram-bot[job-queue]==20.7
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
supabase==2.3.0
asyncpg==0.29.0
